+++ app/services/jobs/return-logs/generate-return-cycle.service.js
@@ -5,7 +5,7 @@
  * @module GenerateReturnCycleService
  */
 
-const { cycleDueDateAsISO, cycleEndDateAsISO, cycleStartDateAsISO } = require('../../../lib/dates.lib.js')
+const { cycleDueDateAsISO, cycleEndDateAsISO, cycleStartDateAsISO } = require('../../../lib/return-cycle-dates.lib.js')
 const ReturnCycleModel = require('../../../models/return-cycle.model.js')
 
 /**
